Separate functions in different files and Manage += errors
