+++ Makefile
@@ -6,7 +6,7 @@
 #    By: rbuitrag <rbuitrag@student.42barcelona.    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/11/06 17:31:12 by yrodrigu          #+#    #+#              #
-#    Updated: 2025/01/28 12:42:32 by yrodrigu         ###   ########.fr        #
+#    Updated: 2025/02/02 13:23:09 by yrodrigu         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -36,6 +36,8 @@ CFILES =	main.c \
 			builtins/env.c\
 			builtins/pwd.c\
 			builtins/export.c\
+			builtins/x_utils.c\
+			builtins/x_utils2.c\
 
 GREEN = "\033[92m"
 RESET = "\033[0m"
