+++ src/builtins/export.c
@@ -6,166 +6,10 @@
 /*   By: yrodrigu <yrodrigu@student.42barcelo>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/28 12:44:57 by yrodrigu          #+#    #+#             */
-/*   Updated: 2025/02/02 13:11:49 by yrodrigu         ###   ########.fr       */
+/*   Updated: 2025/02/02 13:25:24 by yrodrigu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "../../inc/minishell.h"
-
-int	ft_strcmp(char *s1, char *s2)
-{
-	int	i;
-
-	i = 0;
-	while ((s1[i] == s2[i]) && s1[i] && s2[i])
-		i++;
-	return (s1[i] - s2[i]);
-}
-
-void	ft_sort_env(t_env *env)
-{
-	t_env	*front;
-	char	*temp_key;
-	char	*temp_value;
-
-	while (env)
-	{
-		front = env->next;
-		while (front)
-		{
-			if (ft_strcmp(env->key, front->key) > 0)
-			{
-				temp_key = env->key;
-				env->key = front->key;
-				front->key = temp_key;
-				temp_value = env->value;
-				env->value = front->value;
-				front->value = temp_value;
-			}
-			front = front->next;
-		}
-		env = env->next;
-	}
-}
-
-t_env	*ft_find_key_env(t_env *env, char *key_value)
-{
-	t_env	*temp;
-	int		i;
-
-	i = 0;
-	temp = env;
-	while (temp)
-	{
-		if (!ft_strcmp(temp->key, key_value))
-			return (temp);
-		temp = temp->next;
-	}
-	return (NULL);
-}
-
-int	ft_abletojoin(char *str)
-{
-	int	plus;
-	int	i;
-
-	i = 0;
-	plus = 0;
-	while (str[i])
-	{
-		if (str[i] == '+')
-			plus++;
-		i++;
-	}
-	if (plus == 1)
-		return (1);
-	else if (plus > 1)
-		return (2);
-	return (0);
-}
-
-int	ft_key_end(char *str)
-{
-	int	i;
-
-	i = 0;
-	while (str[i])
-	{
-		if (str[i] == '=')
-			return (i);
-		i++;
-	}
-	return (0);
-}
-
-void	ft_check_identifier(char **cmd_array, int *i)
-{
-	while (cmd_array[*i] && cmd_array[*i][0] == '=')
-	{
-		printf("minishell: export: `%s'", cmd_array[*i]);
-		printf(": not a valid identifier\n");
-		(*i)++;
-	}
-}
-
-void	ft_trim_case(char **x_value)
-{
-	char	*temp_x_value;
-
-	temp_x_value = *x_value;
-	printf("trim case before %s\n", *x_value);
-	*x_value = ft_strtrim(*x_value, "\"\'");
-	printf("trim case after %s\n", *x_value);
-	free(temp_x_value);
-}
-
-int	ft_init_key_value(char **cmd_array, char **x_key, char **x_value, int *i)
-{
-	int		end_key;
-	char	*temp_key;
-	int		flag;
-
-	flag = 0;
-	end_key = ft_key_end(cmd_array[*i]);
-	if (end_key)
-	{
-		*x_key = ft_substr(cmd_array[*i], 0, end_key);
-		*x_value = ft_substr(cmd_array[*i], end_key + 1,
-				ft_strlen(cmd_array[*i]));
-	}
-	else
-		*x_key = ft_strdup(cmd_array[*i]);
-	if (ft_abletojoin(*x_key))
-	{
-		temp_key = *x_key;
-		*x_key = ft_strtrim(*x_key, "+");
-		free(temp_key);
-		flag = 1;
-	}
-	ft_trim_case(x_key);
-	ft_trim_case(x_value);
-	return (flag);
-}
-
-void	ft_create_new_node(char **x_key, char **x_value, t_env *env)
-{
-	t_env	*new_node;
-	t_env	*temp;
-
-	new_node = (t_env *)malloc(sizeof(t_env));
-	if (!new_node)
-	{
-		free(*x_key);
-		free(*x_value);
-		return ;
-	}
-	new_node->next = NULL;
-	new_node->key = *x_key;
-	new_node->value = *x_value;
-	temp = env;
-	while (temp->next)
-		temp = temp->next;
-	temp->next = new_node;
-}
 
 void	ft_flag_case1(t_env *node_already_exist, char **x_value)
 {
